refactor: 使用自定義 hook 簡化 VideoTrimmer 元件的狀態管理，並刪除不再使用的類型定義檔，提升代碼可讀性與維護性。
